# gateway-api/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway API
  description: Clinical Data Gateway API
  version: 0.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    nhs-cis2-aal3:
      $ref: https://proxygen.ptl.api.platform.nhs.uk/components/securitySchemes/nhs-cis2-aal3
paths:
  /patient/$gpc.getstructuredrecord:
    post:
      summary: Get structured record
      description: Returns a FHIR Bundle containing patient structured record
      security:
      - nhs-cis2-aal3: []
      operationId: getStructuredRecord
      parameters:
      - in: header
        name: Content-Type
        schema:
          type: string
          enum: [application/fhir+json]
        required: true
      - in: header
        name: Ods-from
        required: true
        schema:
          type: string
          example: "A12345"
          minLength: 1
      - in: header
        name: Ssp-TraceID
        required: true
        schema:
          type: string
          example: "trace-1234"
          minLength: 1
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              required:
                - resourceType
                - parameter
              properties:
                resourceType:
                  type: string
                  enum: ["Parameters"]
                  example: "Parameters"
                parameter:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - name
                      - valueIdentifier
                    properties:
                      name:
                        type: string
                        enum: ["patientNHSNumber"]
                        example: "patientNHSNumber"
                      valueIdentifier:
                        type: object
                        required:
                          - system
                          - value
                        properties:
                          system:
                            type: string
                            minLength: 1
                            example: "https://fhir.nhs.uk/Id/nhs-number"
                          value:
                            type: string
                            pattern: "^[0-9]{10}$"
                            example: "9999999999"
      responses:
        '200':
          description: Successful response
          parameters:
          - in: header
            name: Content-Type
            schema:
              type: string
              enum: [application/fhir+json]
            required: true
          content:
            application/fhir+json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    description: Status code of the interaction
                    example: 200
                  headers:
                    type: object
                    properties:
                      Content-Type:
                        type: string
                        example: "application/fhir+json"
                  body:
                    type: object
                    description: FHIR Bundle containing patient data
                    properties:
                      resourceType:
                        type: string
                        example: "Bundle"
                      id:
                        type: string
                        example: "example-patient-bundle"
                      type:
                        type: string
                        example: "collection"
                      timestamp:
                        type: string
                        format: date-time
                        example: "2026-01-12T10:00:00Z"
                      entry:
                        type: array
                        items:
                          type: object
                          properties:
                            fullUrl:
                              type: string
                              example: "urn:uuid:123e4567-e89b-12d3-a456-426614174000"
                            resource:
                              type: object
                              properties:
                                resourceType:
                                  type: string
                                  example: "Patient"
                                id:
                                  type: string
                                  example: "9999999999"
                                identifier:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      system:
                                        type: string
                                        example: "https://fhir.nhs.uk/Id/nhs-number"
                                      value:
                                        type: string
                                        example: "9999999999"
                                name:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      use:
                                        type: string
                                        example: "official"
                                      family:
                                        type: string
                                        example: "Doe"
                                      given:
                                        type: array
                                        items:
                                          type: string
                                        example: ["John"]
                                gender:
                                  type: string
                                  example: "male"
                                birthDate:
                                  type: string
                                  format: date
                                  example: "1985-04-12"
        '400':
          description: Bad request - invalid input parameters
          content:
            application/fhir+json:
              schema:
                type: object
                properties:
                  resourceType:
                    type: string
                    example: "OperationOutcome"
                  issue:
                    type: array
                    items:
                      type: object
                      properties:
                        severity:
                          type: string
                          example: "error"
                        code:
                          type: string
                          example: "invalid"
                        diagnostics:
                          type: string
                          example: "Invalid NHS number format"
        '404':
          description: Patient not found
          content:
            application/fhir+json:
              schema:
                type: object
                properties:
                  resourceType:
                    type: string
                    example: "OperationOutcome"
                  issue:
                    type: array
                    items:
                      type: object
                      properties:
                        severity:
                          type: string
                          example: "error"
                        code:
                          type: string
                          example: "not-found"
                        diagnostics:
                          type: string
                          example: "Patient not found"
        '500':
          description: Internal server error
          content:
            application/fhir+json:
              schema:
                type: object
                properties:
                  resourceType:
                    type: string
                    example: "OperationOutcome"
                  issue:
                    type: array
                    items:
                      type: object
                      properties:
                        severity:
                          type: string
                          example: "error"
                        code:
                          type: string
                          example: "exception"
                        diagnostics:
                          type: string
                          example: "Internal server error"

        '502':
          description: Received an error response from a downstream server
          content:
            application/fhir+json:
              schema:
                type: object
                properties:
                  resourceType:
                    type: string
                    example: "OperationOutcome"
                  issue:
                    type: array
                    items:
                      type: object
                      properties:
                        severity:
                          type: string
                          example: "error"
                        code:
                          type: string
                          example: "exception"
                        diagnostics:
                          type: string
                          example: "PDS FHIR API request failed: Bad Gateway"